+++ storage/raid/boardcom/storcli/storcli.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+sudo storcli64 /c0 show
+#...
+#FW Package Build = 51.19.0-4234
+#FW Version = 5.190.01-3581