+++ storage/raid/info.sh
@@ -24,7 +24,6 @@ software_raid_info() {
 	[[ -e /dev/md0 ]] && sudo mdadm -D /dev/md0
 }
 
-find_raid
-dev_raid_metadata /dev/sda
-software_raid_info
-
+find_raid || true
+dev_raid_metadata /dev/sda || true
+software_raid_info || true