+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 
 #### Breaking Changes
 #### Added
+- raid: add boardcom/storcli/storcli.sh
 - rootfs: add README.md;
 - virt-install: add cdrom.sh and links;
 - virt-manager: add virt-clone directory;